abbaci

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -10,11 +10,11 @@
 
 int main(void)
 {
-unsigned long long int num1;
-unsigned long long int num2;
-unsigned long long int num3;
-unsigned long long int max;
-unsigned long long int sum;
+unsigned long int num1;
+unsigned long int num2;
+unsigned long int num3;
+unsigned long int max;
+unsigned long int sum;
 num1 = 1;
 num2 = 2;
 max = 4000000;
@@ -29,6 +29,6 @@ if ((num3 % 2) == 0)
 sum += num3;
 }
 }
-printf("%llu\n", sum);
+printf("%lu\n", sum);
 return (0);
 }
